finished implementing the auth of the project using clerk as my auth provider and combined with convex db also i secured all the convex function as good i can now

+++ convex/documents.ts
@@ -1,25 +1,55 @@
 import { ConvexError, v } from "convex/values";
 import { query, mutation } from "./_generated/server";
 
-const userId = "smt";
 export const getById = query({
 	args: { documentID: v.string() },
 	handler: async (ctx, args) => {
+		const identity = await ctx.auth.getUserIdentity();
+
+		if (!identity) {
+			throw new Error("Not authenticated");
+		}
+
+		const userId = identity.subject;
+
 		const normalizedId = ctx.db.normalizeId("documents", args.documentID);
+
 		if (!normalizedId) throw new ConvexError("Invalid document ID");
 		const document = await ctx.db.get(normalizedId);
 
 		if (document == null) {
 			throw new ConvexError("Document couldn't be found");
 		}
+		if (document.user_id !== userId) {
+			throw new Error("Not authorized");
+		}
 
 		return document;
+	},
+});
+
+export const deleteDocumet = mutation({
+	args: { id: v.id("documents") },
+	handler: async (ctx, args) => {
+		const document = await ctx.db.get(args.id);
+		if (!document) throw new Error("Documnet couldnt be found");
+
+		const deletedID = ctx.db.delete(args.id);
+		return deletedID;
 	},
 });
 
 export const getSidebar = query({
 	args: { parentId: v.optional(v.id("documents")) },
 	handler: async (ctx, args) => {
+		const identity = await ctx.auth.getUserIdentity();
+
+		if (!identity) {
+			throw new Error("Not authenticated");
+		}
+
+		const userId = identity.subject;
+
 		const documents = await ctx.db
 			.query("documents")
 			.withIndex("by_user_parent", (q) =>
@@ -32,6 +62,14 @@ export const getSidebar = query({
 });
 export const getUserDocuments = query({
 	handler: async (ctx) => {
+		const identity = await ctx.auth.getUserIdentity();
+
+		if (!identity) {
+			throw new Error("Not authenticated");
+		}
+
+		const userId = identity.subject;
+
 		const documents = await ctx.db
 			.query("documents")
 			.withIndex("by_user_id", (q) => q.eq("user_id", userId))
@@ -44,6 +82,14 @@ export const getUserDocuments = query({
 export const create = mutation({
 	args: { title: v.string(), parentDocument: v.optional(v.id("documents")) },
 	handler: async (ctx, args) => {
+		const identity = await ctx.auth.getUserIdentity();
+
+		if (!identity) {
+			throw new Error("Not authenticated");
+		}
+
+		const userId = identity.subject;
+
 		const document = await ctx.db.insert("documents", {
 			parentDocumet: args.parentDocument,
 			title: args.title,
@@ -62,6 +108,14 @@ export const update = mutation({
 		icon: v.optional(v.string()),
 	},
 	handler: async (ctx, args) => {
+		const identity = await ctx.auth.getUserIdentity();
+
+		if (!identity) {
+			throw new Error("Not authenticated");
+		}
+
+		const userId = identity.subject;
+
 		const { id, ...rest } = args;
 		const normalizedId = ctx.db.normalizeId("documents", args.id);
 		if (!normalizedId) throw new ConvexError("Invalid document ID");
@@ -70,6 +124,7 @@ export const update = mutation({
 		console.log(existingDocs);
 
 		if (!existingDocs) throw new Error("Document couldnt be found");
+		if (existingDocs.user_id !== userId) throw new Error("Not permited");
 
 		await ctx.db.patch(normalizedId, {
 			...rest,
@@ -80,6 +135,18 @@ export const update = mutation({
 export const deleteTask = mutation({
 	args: { id: v.id("documents") },
 	handler: async (ctx, args) => {
+		const identity = await ctx.auth.getUserIdentity();
+
+		if (!identity) {
+			throw new Error("Not authenticated");
+		}
+
+		const userId = identity.subject;
+		const document = await ctx.db.get(args.id);
+
+		if (!document) throw new Error("Document couldnt be found");
+
+		if (document.user_id !== userId) throw new Error("Not authorized");
 		await ctx.db.delete(args.id);
 	},
 });
